refactor: Move non-migrated stories to 'legacy' folder
