+++ .cca/Custom/COMPONENT_NAME.stories.tsx
@@ -3,7 +3,7 @@ import { ComponentStory, Meta } from '@storybook/react';
 import COMPONENT_NAME from './COMPONENT_NAME';
 
 export default {
-  title: 'Components/COMPONENT_NAME',
+  title: 'Legacy/COMPONENT_NAME',
   component: COMPONENT_NAME,
   argTypes: {
     // testProp: {
