+++ .cca/VanillaExtract/src/COMPONENT_NAME.tsx
@@ -1,4 +1,4 @@
-import React, { ReactNode } from 'react';
+import { ReactNode } from 'react';
 import * as styles from './component_name.css';
 // import * as RadixCheckbox from '@radix-ui/react-checkbox';
 
@@ -7,8 +7,6 @@ export type COMPONENT_NAMEProps = {
   children: ReactNode;
 };
 
-const COMPONENT_NAME = ({ children, ...props }: COMPONENT_NAMEProps) => {
-  return <div className={styles.base}>{children}</div>;
+export const COMPONENT_NAME = ({ children, ...props }: COMPONENT_NAMEProps) => {
+  return <div className={styles.component_nameStyle}>{children}</div>;
 };
-
-export default COMPONENT_NAME;
